flask/requirements.txt
Flask==1.1.1
pymongo==3.9.0
dnspython==1.16.0